Move NativeMethods to Aga8.cs file
